# src/docs/swagger.yaml
...
openapi: 3.0.0
info:
  title: joindeveloper - WhatsApp API
  description: |
    </br>
    <hr style="height: 5px;background: #007500;margin: 20px 0;box-shadow: 0px 3px 5px 0px rgb(204 204 204);">

    <div align="center">

    [![Telegram Group](https://img.shields.io/badge/Group-Telegram-%2333C1FF)](https://t.me/joindeveloperBR)
    [![Whatsapp Group](https://img.shields.io/badge/Group-WhatsApp-%2322BC18)](https://chat.whatsapp.com/HyO8X8K0bAo0bfaeW8bhY5)
    [![License](https://img.shields.io/badge/license-GPL--3.0-orange)](https://github.com/code-chat-br/whatsapp-api/blob/main/LICENSE)
    [![Support](https://img.shields.io/badge/Buy%20me-coffe-orange)](https://app.picpay.com/user/cleber.wilson.oliveira)

    </div>
      
    <div align="center"><img src="https://raw.githubusercontent.com/code-chat-br/whatsapp-api/main/public/images/cover.png"></div>

    <font size='3'>
      joindeveloper is an advanced WhatsApp integration API that seamlessly merges the popularity of WhatsApp with modern business applications. With comprehensive features, the joindeveloper API empowers businesses to expand their digital presence, reach customers more directly, and provide personalized and efficient support, all through the world's most widely used messaging platform.

...
    </font>
  
    [![Run in Postman](https://run.pstmn.io/button.svg)](https://www.postman.com/joindeveloper/workspace/joindeveloper-whatsapp-api/api/fbe06c7b-7647-4c71-81ee-841f5b2e90d8?action=share&creator=14064846)
...
servers: []
components:
  securitySchemes:
    apikeyAuth:
      type: apiKey
      in: header
      name: apikey
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []
...
paths:
...
  /instance/create:
    post:
      tags:
        - Instance Controller
      summary: Create Instance
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                instanceName:
                  type: string
                  description: Name of the instance (optional).
              example:
                instanceName: 'exampleInstance'
      security:
        - apikeyAuth: []
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
  /instance/fetchInstances:
    get:
      tags:
        - Instance Controller
      summary: Fetch Instances
      security:
        - apikeyAuth: []
      parameters:
        - name: instanceName
          in: query
          schema:
            type: string
          description: Retrieve one or all instances (optional).
          example: 'joindeveloper'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items: 
                  type: object
                  properties:
                    instance:
                      type: object
                      properties:
                        instanceName:
                          type: string
                        owner:
                          type: string
                        profileName:
                          type: string
                        profilePictureUrl:
                          type: string
                    auth:
                      type: object
                      properties:
                        jwt:
                          type: string
...
  /instance/refreshToken/:
    put:
      tags:
        - JWT
      summary: Refresh an expired JWT token.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                oldToken:
                  type: string
                  description: '- required - The expired JWT token.'
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
...